Extreme Computing/outputcrap/EXCbackup/task7/requirements.txt
Used python libraries:

bitarray-0.8.1
mmh3-2.5.1
